Original game plus source.
This is a raw commit of the original Dudes of Stuff and Things, binaries and
all, along with the original source.  The only modification done was to make
all of the files with lowercase names; since they originally came from DOS,
case is irrelevant.  No files have been otherwise edited from the original
ZIP file.